django_todo/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
PyJWT
